pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "MLOps lab 1"
authors = ["Игорь Иванов <is.ivanov@bars.group>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
flask-cors = "^4.0.0"
boto3 = "^1.29.3"
matplotlib = "^3.8.2"
torch = "^2.1.1"
pyyaml = "^6.0.1"
torchvision = "^0.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
